ShadowCameraManager: do not override openCameraDeviceUserAsync on V+
Attempting to set the platform codename to Baklava will break
robolectric:

    external/robolectric/shadows/framework/src/main/java/org/robolectric/shadows/ShadowCameraManager.java:128: error: No method openCameraDeviceUserAsync(java.lang
    .String,android.hardware.camera2.CameraDevice.StateCallback,java.util.concurrent.Executor,int,int,int) in android.hardware.camera2.CameraManager for SDK 10000
      protected CameraDevice openCameraDeviceUserAsync(

In the current codebase, there is no method

    openCameraDeviceUserAsync(String, CameraDevice.StateCallback, Executor, int, int, int)

(but there is one with one less int argument).

Unblock the work to set the codename to Baklava by limiting the 3 int
robolectric shadow to V only.

Bug: 354023031
Test: treehugger
Ignore-AOSP-First: unblock work to set codename to Baklava (this CL can be CP'd to AOSP afterwards)
Change-Id: Id77eb904f04ae6ce9f60cb0c98db2a6bacee104e